fix(web): remove VITE_ERROR_TRACKING_URL and split non-component exports
